requirements.txt
streamlit
openai
pymongo
pandas
azure-storage-blob
pytesseract
Pillow
moviepy
pytube
openai-whisper